Conflicts:
	DrupalCon/DrupalCon.xcodeproj/project.pbxproj
	DrupalCon/DrupalCon.xcodeproj/project.xcworkspace/xcshareddata/DrupalCon.xccheckout
	DrupalCon/DrupalCon/Classes/ConferenceViewControllers/Program/DCProgramViewController.m
	DrupalCon/DrupalCon/Classes/ConferenceViewControllers/Speakers/DCSpeakersViewController.m
	DrupalCon/DrupalCon/Classes/Services/DCMainProxy+Additions.m
	DrupalCon/DrupalCon/Classes/UIConstants/UIConstants.h
	DrupalCon/DrupalCon/Model/Categories/DCEvent+DC.m
	DrupalCon/DrupalCon/Model/Categories/DCSpeaker+DC.m
	DrupalCon/DrupalCon/Model/DCMainProxy.m
	DrupalCon/DrupalCon/Model/main.xcdatamodeld/main.xcdatamodel/contents
	DrupalCon/DrupalCon/Resources/Storyboards/Main.storyboard

+++ DrupalCon/DrupalCon/Classes/ConferenceViewControllers/Program/DCProgramItemsViewController.m
@@ -243,11 +243,9 @@ static NSString *const cellIdSpeechOfDay = @"ProgramCellIdentifierSpeechOfDay";
     DCEvent * event = [self eventForIndexPath:anIndexPath];
     event.favorite = [NSNumber numberWithBool:isFavorite];
     if (isFavorite) {
-        [[DCMainProxy sharedProxy]
-         addToFavoriteEvent:event];
+        [[DCMainProxy sharedProxy] addToFavoriteEvent:event];
     } else {
-        [[DCMainProxy sharedProxy]
-         removeFavoriteEventWithID:event.eventID];
+        [[DCMainProxy sharedProxy] removeFavoriteEventWithID:event.eventID];
     }
 }
 
@@ -369,7 +367,7 @@ static NSString *const cellIdSpeechOfDay = @"ProgramCellIdentifierSpeechOfDay";
 
 - (BOOL)isClickEnableForEvent:(DCEvent *)event
 {
-    int type = [event getTypeID];
+    NSInteger type = [event getTypeID];
     return !(type != DC_EVENT_GROUP &&
             type != DC_EVENT_REGISTRATION &&
             type != DC_EVENT_LUNCH);
